requirements.txt
amqp==2.3.2
billiard==3.5.0.4
boto3==1.7.48
celery==4.2.1
certifi==2018.1.18
dj-database-url==0.4.2
Django==1.10.8
git+https://github.com/venicegeo/django-audit-logging.git@v0.1.8#egg=django-audit-logging
django-auth-ldap==1.2.12
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-extensions==1.7.7
django-filter==1.0.1
django-notifications-hq==1.3
djangorestframework==3.6.0
djangorestframework-gis==0.11
enum34==1.1.6
eventlet==0.20.1
GDAL==2.2.2
gunicorn==19.7.1
# Kombu declared explicitly because of incompatibility with celery requirements.
kombu==4.2.1
lxml==3.7.3
MapProxy==1.11.0
Markdown==2.6.8
numpy==1.14.3
Pillow==4.1.1
psycopg2==2.7.3.2
PyJWT==1.4.2
pyparsing==2.1.10
pysqlite==2.8.3
python-dateutil==2.7.2
python-ldap==2.4.39
python-magic==0.4.12
pytz==2017.2
PyYAML==3.12
requests==2.18.4
Shapely==1.5.17
simplejson==3.10.0
whitenoise==3.3.0
